[REFACTOR] properties 설정 파일들 config/properties 로 변경
