+++ src/main/java/store/ckin/front/deliverypolicy/adapter/impl/DeliveryPolicyAdapterImpl.java
@@ -10,7 +10,7 @@ import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
-import store.ckin.front.config.GatewayProperties;
+import store.ckin.front.config.properties.GatewayProperties;
 import store.ckin.front.deliverypolicy.adapter.DeliveryPolicyAdapter;
 import store.ckin.front.deliverypolicy.dto.request.DeliveryPolicyCreateRequestDto;
 import store.ckin.front.deliverypolicy.dto.request.DeliveryPolicyUpdateRequestDto;
