+++ src/main/java/store/ckin/front/pointpolicy/adapter/impl/PointPolicyAdapterImpl.java
@@ -10,7 +10,7 @@ import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
-import store.ckin.front.config.GatewayProperties;
+import store.ckin.front.config.properties.GatewayProperties;
 import store.ckin.front.pointpolicy.adapter.PointPolicyAdapter;
 import store.ckin.front.pointpolicy.dto.request.PointPolicyCreateRequestDto;
 import store.ckin.front.pointpolicy.dto.request.PointPolicyUpdateRequestDto;
